Added Python 3 parellel matrix multiply
The Python matrix mulitply is called by making the script itself
exectuable and then calling it with mpirun -np 7 (name of script) (name
of file to read for the matricies). Currently, I need to add timers and
test on the IU clusters but for now it ran fine on I3 laptop running
Gallium. It uses the master-slave concept as in it uses the strassen
algorthm to pass around all information regarding the original matrices
and then uses numpy to perform calculations on that. Attempted to remain
as simple as possibly and use Numpy as much as possible as numpy uses C
for the calculations and is optimized better than I can do anything.